Separate header from actual implementation
